* Calculate practice information (kcal, duration) * Using databinding to make the code shorter * Update buttons status * handle pause/resume practice

+++ RunStatistic/app/src/main/java/com/tantnt/android/runstatistic/models/PracticeModel.kt
@@ -16,7 +16,7 @@ import com.tantnt.android.runstatistic.database.DatabasePractice
 
 data class PracticeModel(
     var start_time : Long,
-    var duration: Long,
+    var duration: Double,
     var distance: Double,
     var calo: Double,
     var speed: Double,
